feat: migrate to bigint

+++ src/inclusion.js
@@ -4,6 +4,7 @@ import { maxIndexEntriesInDeal, EntrySize } from './index.js'
 /**
  *
  * @param {API.PaddedPieceSize} size
+ * @returns {bigint}
  */
 export const indexAreaStart = (size) =>
-  size - maxIndexEntriesInDeal(size) * EntrySize
+  size - BigInt(maxIndexEntriesInDeal(size) * EntrySize)
